Refactor : LottoResultTest 클래스 LottoRankTest 클래스로 이름 변경
